Functionality complete for food journal and food items

+++ src/main/java/com/revature/nutritioknights/foodentry/FoodEntryService.java
@@ -5,13 +5,13 @@ import com.revature.nutritioknights.foodentry.dtos.GetByDateRequest;
 import com.revature.nutritioknights.foodentry.dtos.GetByMealnameRequest;
 import com.revature.nutritioknights.foodentry.dtos.NewFoodEntryRequest;
 import com.revature.nutritioknights.util.annotations.Inject;
+import com.revature.nutritioknights.util.custom_exceptions.InvalidRequestException;
+import com.revature.nutritioknights.util.custom_exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.util.Date;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 @Service
 @Transactional
@@ -23,29 +23,59 @@ public class FoodEntryService {
     @Autowired
     public FoodEntryService (FoodEntryRepository foodEntryRepository){
         this.foodEntryRepository=foodEntryRepository;
-        //we don't actually persist any food items, only their ids.
-
     }
 
     public String newEntry(NewFoodEntryRequest request){
         FoodEntry newFoodEntry = new FoodEntry(request);
         newFoodEntry.setDateInt(new Date().getTime()/(1000*60*60*24));
         newFoodEntry.setEntry_id(UUID.randomUUID().toString());
+        //input validation
+        if (newFoodEntry.getServing_amt()<0){
+            throw new InvalidRequestException ("You can't have negative serving sizes, unfortunately.");
+        }
+        if (newFoodEntry.getServing_amt()==0){
+            throw new InvalidRequestException ("Having 0 servings of something is the same as not eating it, silly!");
+        }
         foodEntryRepository.save(newFoodEntry);
         return newFoodEntry.getEntry_id();
     }
 
     public List<FoodEntry> getUserEntriesByDate(GetByDateRequest request){
-        return foodEntryRepository.getAllByDateIntAndUsername(request.getDateInt(),request.getUsername());
+        List<FoodEntry> entryList = foodEntryRepository.getAllByDateIntAndUsername(request.getDateInt(),request.getUsername());
+        if (entryList.size()==0){
+            throw new NotFoundException("No entries found for this combination of user and date.");
+        }
+        return entryList;
     }
 
     public List<FoodEntry> getUserEntriesByMealname(GetByMealnameRequest request){
-        return foodEntryRepository.getAllByMealnameIdAndUsername(request.getMealnameId(),request.getUsername());
+        List<FoodEntry> entryList = foodEntryRepository.getAllByMealnameIdAndUsername(request.getMealnameId(),request.getUsername());
+        if (request.getMealnameId()>6||request.getMealnameId()<1) {
+            throw new InvalidRequestException("Invalid mealname id (must be 1-6).");
+        }
+        if (entryList.size()==0){
+            throw new NotFoundException("No entries found for the selected combination of meal and date.");
+        }
+        return entryList;
     }
     public List<Long> getActivity(String username){
-        return foodEntryRepository.getDateintsByUsername(username);
+        List<Long> entryList= foodEntryRepository.getDateintsByUsername(username);
+        if (entryList.size()==0){
+            return null;
+        }
+        return entryList;
     }
     public void deleteEntry(String id){
+        if (!(this.getById(id).getDateInt()==new Date().getTime()/(1000*60*60*24))) {
+            throw new InvalidRequestException("You cannot delete entries from a previous date.");
+        }
         foodEntryRepository.deleteById(id);
+    }
+    public FoodEntry getById(String id){
+        try {
+            return foodEntryRepository.findById(id).get();
+        } catch (NoSuchElementException e){
+            throw new NotFoundException("Could not find an entry with that ID");
+        }
     }
 }
